refactor(config): extract llm-config-map.ts (R2)
